+++ src/main/Java/dk/aau/cs/giraf/tortoise/controller/AbstractSequence.java
@@ -4,7 +4,7 @@ package dk.aau.cs.giraf.tortoise.controller;
 public abstract class AbstractSequence {
 
     private long id;
-	protected int titlePictoId;
+	protected long titlePictoId;
 	protected String title;
 	protected int numChoices;
 	
@@ -14,7 +14,7 @@ public abstract class AbstractSequence {
 		title = "";
 	}
 
-    public AbstractSequence(long id, int titlePictoId, String title){
+    public AbstractSequence(long id, long titlePictoId, String title){
         this.id = id;
         this.titlePictoId = titlePictoId;
         this.title = title;
@@ -25,15 +25,15 @@ public abstract class AbstractSequence {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 	
-	public int getTitlePictoId() {
+	public long getTitlePictoId() {
 		return titlePictoId;
 	}
 
-	public void setTitlePictoId(int titlePictoId) {
+	public void setTitlePictoId(long titlePictoId) {
 		this.titlePictoId = titlePictoId;
 	}
